lesson22
lesson22

+++ blog/app/Controller/PostsController.php
@@ -43,10 +43,17 @@ class PostsController extends AppController {
             throw new MethodNotAllowedException();
             
         }
-        if($this->Post->delete($id)) {
-            $this->Session->setFlash('Deleted!');
-            $this->redirect(array('action'=>'index'));
+        if ($this->request->is('ajax')) {
+            if ($this->Post->delete($id)) {
+                $this->autoRender = false;
+                $this->autoLayout = false;
+                $response = array('id' => $id);
+                $this->header('Content-Type: application/json');
+                echo json_encode($response);
+                exit();
+            }
         }
+        $this->redirect(array('action'=>'index'));
     }
 }
